Generalized qe-test to handle multiple runs in a single test.
Each test directory, e.g. "Na" or "Gr-hex", should contain a
set of 'runs', e.g. "ref" or "test", each in their own self-named
subdirectories.  The 'ref' run is treated specially: it is copied
to 'old' and used to validate fall-back behavior.  The other runs
are executed and timed.

Right now, the output of the 'ref' and 'test' runs are compared as
before.  In the next commit, the output will be formatted as a
table with runs as columns.  This can be seen in the display of
runtimes.

qe-run.py had its -e option re-implemented to specify the location
holding qe-related binaries.