First pass at adding VASP support (doesn't actually deal with VASP).
The big issue is that VASP and QE use very different output files.
One approach would be to turn each output file into a universal
data structure, that could then be read.  I think this solution
is too heavy for now.

Instead, the patterns which are used to pick out the output values
are put in a dict with keys corresponding to output formats.  In
each run directory, a loop looks for types of output files and
uses the appropriate sets of patterns.

This commit makes the above generalization and works for total
energy and pressure in qe runs.

In the future, this pattern <=> output file association could be
generalized to a function <=> output file generalization, which
might be necessary for reading some more complicated things.